+++ tests/test_testing/test_builder.py
@@ -1,18 +1,22 @@
 import uuid
+from datetime import datetime
 from io import StringIO
 
 import pytest
 from pydantic import ValidationError
 
 from botx import (
+    ChatTypes,
     Entity,
     EntityTypes,
     File,
+    Forward,
     Mention,
     MentionTypes,
     MessageBuilder,
     UserMention,
 )
+from botx.models.entities import Reply
 
 
 def test_setting_new_file_from_file():
@@ -63,8 +67,8 @@ def test_mention_user_in_message():
     builder = MessageBuilder()
     builder.mention_user(user_huid)
 
-    assert builder.message.entities[0].data.mention_type == MentionTypes.user
-    assert builder.message.entities[0].data.mention_data.user_huid == user_huid
+    assert builder.message.entities.__root__[0].data.mention_type == MentionTypes.user
+    assert builder.message.entities.__root__[0].data.mention_data.user_huid == user_huid
 
 
 def test_mention_contact_in_message():
@@ -72,8 +76,10 @@ def test_mention_contact_in_message():
     builder = MessageBuilder()
     builder.mention_contact(user_huid)
 
-    assert builder.message.entities[0].data.mention_type == MentionTypes.contact
-    assert builder.message.entities[0].data.mention_data.user_huid == user_huid
+    assert (
+        builder.message.entities.__root__[0].data.mention_type == MentionTypes.contact
+    )
+    assert builder.message.entities.__root__[0].data.mention_data.user_huid == user_huid
 
 
 def test_mention_chat_in_message():
@@ -81,8 +87,10 @@ def test_mention_chat_in_message():
     builder = MessageBuilder()
     builder.mention_chat(chat_id)
 
-    assert builder.message.entities[0].data.mention_type == MentionTypes.chat
-    assert builder.message.entities[0].data.mention_data.group_chat_id == chat_id
+    assert builder.message.entities.__root__[0].data.mention_type == MentionTypes.chat
+    assert (
+        builder.message.entities.__root__[0].data.mention_data.group_chat_id == chat_id
+    )
 
 
 def test_setting_raw_entities():
@@ -94,7 +102,7 @@ def test_setting_raw_entities():
         ),
     ]
 
-    assert builder.message.entities[0].data.mention_type == MentionTypes.user
+    assert builder.message.entities.__root__[0].data.mention_type == MentionTypes.user
 
 
 @pytest.mark.parametrize(
@@ -124,3 +132,65 @@ def test_error_when_file_validation_not_passed():
     builder.body = "file_transfer"
     with pytest.raises(ValidationError):
         builder.system_command = True
+
+
+class TestBuildForward:
+    def test_building_forward_via_models(self):
+        builder = MessageBuilder()
+        builder.forward(
+            forward=Forward(
+                group_chat_id=uuid.uuid4(),
+                sender_huid=uuid.uuid4(),
+                forward_type=ChatTypes.group_chat,  # ignore: type
+                source_inserted_at=datetime.now(),
+            ),
+        )
+
+    def test_building_forward_via_message(self, message):
+        builder = MessageBuilder()
+        builder.forward(message=message)
+
+    def test_building_forward_arguments_error(self, message):
+        builder = MessageBuilder()
+        with pytest.raises(ValueError):
+            builder.forward(
+                message=message,
+                forward=Forward(
+                    group_chat_id=uuid.uuid4(),
+                    sender_huid=uuid.uuid4(),
+                    forward_type=ChatTypes.botx,  # ignore: type
+                    source_inserted_at=datetime.now(),
+                ),
+            )
+
+
+class TestBuildReply:
+    def test_building_forward_via_models(self):
+        builder = MessageBuilder()
+        builder.reply(
+            reply=Reply(
+                body="foo",
+                reply_type=ChatTypes.group_chat,
+                sender=uuid.uuid4(),
+                source_chat_name="bar",
+                source_sync_id=uuid.uuid4(),
+            ),
+        )
+
+    def test_building_forward_via_message(self, message):
+        builder = MessageBuilder()
+        builder.reply(message=message)
+
+    def test_building_forward_arguments_error(self, message):
+        builder = MessageBuilder()
+        with pytest.raises(ValueError):
+            builder.reply(
+                message=message,
+                reply=Reply(
+                    body="foo",
+                    reply_type=ChatTypes.group_chat,
+                    sender=uuid.uuid4(),
+                    source_chat_name="bar",
+                    source_sync_id=uuid.uuid4(),
+                ),
+            )